+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to `@homebridge-plugins/homebridge-matter` will be documente
 ### Changed
 
 - feat: allow the thermostat's auto, heat and cool modes to be removed in the config
+- feat: allow the smoke and co alarms to be removed in the config
 
 ## v1.6.0 (2026-08-09)